Revert "Move channel and codec related files to common (#4262)" (#4277)
This reverts commit a48b5f6d25c7371d05d6d8290e385b83d58f0b72.
