+++ shell/platform/darwin/common/BUILD.gn
@@ -9,14 +9,6 @@ source_set("common") {
   sources = [
     "buffer_conversions.h",
     "buffer_conversions.mm",
-    "FlutterBinaryMessenger.h",
-    "FlutterChannels.h",
-    "FlutterChannels.mm",
-    "FlutterCodecs.h",
-    "FlutterCodecs.mm",
-    "FlutterMacros.h",
-    "FlutterStandardCodec.mm",
-    "FlutterStandardCodec_Internal.h",
     "platform_mac.h",
     "platform_mac.mm",
     "process_info_mac.cc",
