+++ shell/platform/darwin/ios/framework/Source/FlutterPlatformPlugin.h
@@ -5,7 +5,7 @@
 #ifndef SHELL_PLATFORM_IOS_FRAMEWORK_SOURCE_FLUTTERPLATFORMPLUGIN_H_
 #define SHELL_PLATFORM_IOS_FRAMEWORK_SOURCE_FLUTTERPLATFORMPLUGIN_H_
 
-#include "flutter/shell/platform/darwin/common/FlutterChannels.h"
+#include "flutter/shell/platform/darwin/ios/framework/Headers/FlutterChannels.h"
 
 @interface FlutterPlatformPlugin : NSObject
