+++ shell/platform/darwin/ios/framework/Source/flutter_standard_codec_unittest.mm
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#include "flutter/shell/platform/darwin/common/FlutterCodecs.h"
+#include "flutter/shell/platform/darwin/ios/framework/Headers/FlutterCodecs.h"
 #include "gtest/gtest.h"
 
 void checkEncodeDecode(id value, NSData* expectedEncoding) {
